all: convert flowcontrol to TypeScript
