+++ src/flowcontrol/Fn.ts
@@ -1,0 +1,2 @@
+export type Fn = (...args: any[]) => void;
+export type CancelableFn = Fn | { cancel: () => void };
